feat: allow declaration of the filter in the Campaign frontmatter
Depend on the use case.

One can create a campaign for a subset of the recipients.
In this case, having the filter is the campaign is the most logic.

Other can plan to send the campaign to all the recipients. But for technical reason, it make sense to send by batches.
In this cas, having the filter on the command line make more sense.